Introduction
Depression is a psychological disorder that affects over three hundred million humans worldwide. A person who is depressed suffers from anxiety in day-to-day life, which affects that person in the relationship with their family and friends, leading to different diseases and in the worst-case death by suicide. With the growth of the social network, most people share their emotion, feelings, their thoughts on social media. If their depression can be detected early by analyzing their post, then by taking the necessary steps, a person can be saved from depression-related diseases, or in the best case he can be saved from committing suicide. In this research work, a hybrid model has been proposed that can detect depression by analyzing users’ textual posts. Deep learning algorithms were trained using the training data and then performance has been evaluated on the test data of the dataset of Reddit which was published for the pilot piece of work.

The World Health Organization (WHO) predicts that depression disorders will be widespread in the next 20 years. We try to identify the most effective deep neural network architecture among a few selected architectures that were successfully used in natural language processing tasks. The proposed method employs Recurrent Neural Networks to compute the post representation of each individual. The representations are then combined with other content-based, behavior and living environment features to predict the depression label of the individual with Deep Neural Networks.